Centralize interfaces
We're going to test out using ChatGPT to generate interfaces, and having
all the props will make this better.
