+++ packages/ui/src/Card.tsx
@@ -3,21 +3,19 @@ import React from "react";
 import {Box} from "./Box";
 import {BoxProps} from "./Common";
 
-export class Card extends React.Component<BoxProps, {}> {
-  render() {
-    return (
-      <Box
-        color={this.props.color || "white"}
-        direction="column"
-        display="flex"
-        padding={this.props.padding || 4}
-        rounding={3}
-        shadow
-        width={this.props.width}
-        {...this.props}
-      >
-        {this.props.children}
-      </Box>
-    );
-  }
+export function Card({children, color = "white", padding = 4, width, ...rest}: BoxProps) {
+  return (
+    <Box
+      color={color}
+      direction="column"
+      display="flex"
+      padding={padding}
+      rounding={3}
+      shadow
+      width={width}
+      {...rest}
+    >
+      {children}
+    </Box>
+  );
 }
