+++ packages/ui/src/InfoTooltipButton.tsx
@@ -1,12 +1,7 @@
 import React from "react";
 
-import {IconSize} from "./Common";
+import {InfoTooltipButtonProps} from "./Common";
 import {IconButton} from "./IconButton";
-
-interface InfoTooltipButtonProps {
-  text: string;
-  size?: IconSize;
-}
 
 export function InfoTooltipButton({text, size}: InfoTooltipButtonProps): React.ReactElement {
   return (
