--- packages/ui/src/Layer.tsx
@@ -1,17 +1,0 @@
-import React from "react";
-
-import {LayerProps} from "./Common";
-
-interface LayerState {}
-
-// TODO: Flesh out for native.
-export class Layer extends React.Component<LayerProps, LayerState> {
-  constructor(props: LayerProps) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    return this.props.children;
-  }
-}
