+++ packages/ui/src/Link.tsx
@@ -1,11 +1,8 @@
 import React from "react";
 import {Linking} from "react-native";
 
-import {Text, TextProps} from "./Text";
-
-interface LinkProps extends TextProps {
-  href: string;
-}
+import {LinkProps} from "./Common";
+import {Text} from "./Text";
 
 export function Link(props: LinkProps): React.ReactElement {
   return (
