--- packages/ui/src/Meta.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import {MetaProps} from "./Common";
-
-export class Meta extends React.Component<MetaProps, {}> {
-  render() {
-    return this.props.children || null;
-  }
-}
