+++ packages/ui/src/SideDrawer.tsx
@@ -2,23 +2,8 @@ import React, {ReactElement} from "react";
 import {Platform, SafeAreaView, StyleProp, ViewStyle} from "react-native";
 import {Drawer} from "react-native-drawer-layout";
 
+import {SideDrawerProps} from "./Common";
 import {isMobileDevice} from "./MediaQuery";
-
-export interface SideDrawerProps {
-  // Position of the drawer relative to the child
-  position?: "right" | "left";
-  // Used to open/hide drawer. Use the onClose and onOpen props to control state
-  isOpen: boolean;
-  // Content within the drawer
-  renderContent: () => ReactElement | ReactElement[];
-  // TODO: Allow the hardware back button on Android to close the SideDrawer
-  onClose?: () => void;
-  onOpen?: () => void;
-  drawerType?: "front" | "back" | "slide" | "permanent";
-  // Content that is wrapped by the drawer. The drawer will use the height of the child it wraps. Can be overwritten via styles prop
-  children?: ReactElement;
-  drawerStyles?: StyleProp<ViewStyle>;
-}
 
 const DEFAULT_STYLES: StyleProp<ViewStyle> = {
   width: Platform.OS === "web" ? "40%" : "95%",
