+++ packages/ui/src/Switch.tsx
@@ -4,16 +4,10 @@ import {Switch as NativeSwitch} from "react-native";
 import {SwitchProps} from "./Common";
 import {WithLabel} from "./WithLabel";
 
-export class Switch extends React.Component<SwitchProps, {}> {
-  render() {
-    return (
-      <WithLabel labelAlignItems="center" labelInline labelJustifyContent="between" {...this.props}>
-        <NativeSwitch
-          disabled={this.props.disabled}
-          value={this.props.switched}
-          onValueChange={this.props.onChange}
-        />
-      </WithLabel>
-    );
-  }
+export function Switch({disabled, switched, onChange, ...rest}: SwitchProps): React.ReactElement {
+  return (
+    <WithLabel labelAlignItems="center" labelInline labelJustifyContent="between" {...rest}>
+      <NativeSwitch disabled={disabled} value={switched} onValueChange={onChange} />
+    </WithLabel>
+  );
 }
