+++ packages/ui/src/index.tsx
@@ -30,7 +30,6 @@ export * from "./InfoTooltipButton";
 export * from "./Link";
 export * from "./Mask";
 export * from "./MediaQuery";
-export * from "./Meta";
 export * from "./MobileAddressAutoComplete";
 export * from "./Modal";
 export * from "./ModalSheet";
@@ -57,7 +56,6 @@ export * from "./Theme";
 export * from "./Toast";
 export * from "./Tooltip";
 export * from "./UnifiedAddressAutoComplete";
-export * from "./UnifiedScreens";
 export * from "./Unifier";
 export * from "./useStoredState";
 export * from "./WebAddressAutocomplete";
